refactor color selector
